+++ docs/gallery/gallery/customization/legend.jl
@@ -45,3 +45,21 @@ grid = draw!(ax_scatter, plt)
 legend!(makie_fig[1, 1], grid; tellheight=false, tellwidth=false, halign=:right, valign=:top)
 
 makie_fig
+
+# If the automatic legend elements are not legible enough, you can change their properties
+# by passing overrides to the `legend` attribute of a `visual`.
+
+df = (;
+    x = repeat(1:100, 5),
+    y = reduce(vcat, [[cos(x) for x in range(0, 8pi, length = 100)] .+ 0.3 .* randn.() for _ in 1:5]),
+    group = repeat(1:5, inner = 100),
+)
+
+lin = data(df) *
+    mapping(:x, :y, group = :group => nonnumeric) *
+    visual(Lines, linewidth = 0.3, label = "Lines", legend = (; linewidth = 1.5))
+sca = data(df) *
+    mapping(:x, :y => y -> y + 5, group = :group => nonnumeric) *
+    visual(Scatter, markersize = 3, label = "Scatter", legend = (; markersize = 12))
+
+draw(lin + sca)