+++ src/guides/legend.jl
@@ -194,7 +194,7 @@ function compute_legend(grid::Matrix{AxisEntries}; order::Union{Nothing,Abstract
                 push!(used_scales, :Label)
                 for (label, processedlayers) in pairs(labelled_layers)
                     push!(legend_els, mapreduce(vcat, processedlayers) do p
-                        legend_elements(p, MixedArguments())
+                        _legend_elements(p, MixedArguments())
                     end)
                     push!(datalabs, label)
                 end
@@ -275,7 +275,7 @@ function compute_legend(grid::Matrix{AxisEntries}; order::Union{Nothing,Abstract
                         # skip the legend element for this processed layer if the kwargs are empty
                         # which means that no scale in this merge group affected this processedlayer
                         if !isempty(kwargs)
-                            append!(_legend_els[i], legend_elements(processedlayer, kwargs))
+                            append!(_legend_els[i], _legend_elements(processedlayer, kwargs))
                         end
                     end
                 end
@@ -306,14 +306,39 @@ function datavalues_plotvalues_datalabels(aes::Type{AesMarkerSize}, scale::Conti
     datavalues, markersizes, string.(datavalues)
 end
 
+function _legend_elements(processedlayer, scale_args::MixedArguments)
+    els = legend_elements(processedlayer, scale_args)
+    if haskey(processedlayer.attributes, :legend)
+        # LegendOverride and apply_legend_override are not public API (introduced in 0.21.13)
+        # but unlikely to change much so it's easier to just depend on them instead of copying
+        # their implementation
+        override = Makie.LegendOverride(processedlayer.attributes[:legend])
+        for el in els
+            Makie.apply_legend_override!(el, override)
+        end
+    end
+    return els
+end
+
 function legend_elements(p::ProcessedLayer, scale_args::MixedArguments)
-    legend_elements(p.plottype, p.attributes, scale_args)
+    els = legend_elements(p.plottype, p.attributes, scale_args)
+    for el in els
+        el.attributes[:alpha] = _get(p.plottype, p.attributes, scale_args, :alpha, 1.0)
+    end
+    return els
 end
 
 function _get(plottype, scale_args, attributes, key)
     get(scale_args, key) do
         get(attributes, key) do 
             to_value(Makie.default_theme(nothing, plottype)[key])
+        end
+    end
+end
+function _get(plottype, scale_args, attributes, key, fallback)
+    get(scale_args, key) do
+        get(attributes, key) do
+            to_value(get(Makie.default_theme(nothing, plottype), key, fallback))
         end
     end
 end
@@ -370,7 +395,7 @@ function legend_elements(T::Type{Heatmap}, attributes, scale_args::MixedArgument
     )]
 end
 
-function legend_elements(T::Type{<:Union{HLines,VLines,Lines,LineSegments,Errorbars,Rangebars,Wireframe,ABLines,ECDFPlot}}, attributes, scale_args::MixedArguments)
+function legend_elements(T::Type{<:Union{HLines,VLines,Lines,LineSegments,Errorbars,Rangebars,Wireframe,ABLines,ECDFPlot,Stairs}}, attributes, scale_args::MixedArguments)
 
     is_vertical = T === VLines || (T <: Union{Errorbars,Rangebars} && _get(T, scale_args, attributes, :direction) === :y)
     # TODO: seems errorbars and rangebars are missing linestyle in Makie, once this is fixed, remove this